Edit cash out tests (#119)
* moved test_cash_out_accuracy from omnipool_amm to omnipool_state

* in test_cash_out_accuracy, give agent shares and holdings in every asset
